Added a Title screen
Added a title screen, that has a working Play, Options, and quit button.

The Options panel has the ability to request the users computers resoultion, and create options for them to work. There is also a full screen, and graphics option button, which also do work. Finally their is a working volume slider, that does change the master volume settings.