Homebase Start finished.
The user is now able to go up to a object (Needs to change), and is able to click 'E' on it so that it transfers them to a new scene so they can start their adventure. The user is also able to return back home by going to the object (Also, needs to change), and clicking 'E' to return home.